Moved characterCreateScreen to its own package Added tabContent methods to own files
